Extract interface from BuildCacheCommandFactory
